Rename: AlertState.tsx into state.ts
