+++ src/components/Alert/state.ts
